+++ dotnet/samples/Concepts/AgentSyntax/OpenAIAssistant/OpenAIAssistant_Agent.cs
