+++ dotnet/samples/Concepts/AgentSyntax/OpenAIAssistant/OpenAIAssistant_ChartMaker.cs
