+++ dotnet/samples/Concepts/AgentSyntax/OpenAIAssistant/OpenAIAssistant_Retrieval.cs
